Model_desktop/requirements.txt

IMDbPY==2022.7.9
keyboard==0.13.5
Kivy==2.2.1
numpy==1.26.3
PyAutoGUI==0.9.54
python-decouple==3.8
pyttsx3==2.90
pywhatkit==5.4
Requests==2.31.0
sounddevice==0.4.6
SpeechRecognition==3.10.0
wikipedia==1.4.0
wolframalpha==5.0.0
